Add files via upload
The aim of this Regression project is to use machine learning techniques to predict the productivity of garment workers based on various factors such as the type of product, the number of workers, and the target quantity. The productivity is measured by the ratio of the actual output to the expected output.
The project involves the following steps:
Data preprocessing: This includes outlier detection and removal, encoding of categorical variables, and scaling of numerical variables.
Model selection: This involves choosing and comparing different machine learning algorithms such as decision tree, random forest, support vector machine, etc. to find the best model for the regression task.
Model evaluation: This involves measuring the performance of the selected model using metrics such as mean squared error, and  mean absolute error.
The project aims to provide insights into the factors that affect the productivity of garment workers and help the management to improve the efficiency and quality of the production process.